TST: use rel imports from tests
